code cleanup and xlsx file isssue solved

+++ app/routers/ai_models.py
@@ -4,37 +4,50 @@ from fastapi import APIRouter, Depends, HTTPException, UploadFile
 from app.data_deletion import delete_collection
 from app.file_reader import read_file_return_csv
 from app.schemas import session_schemas as schemas
+from app.schemas import ai_assistant_schemas as ai_schemas
 from app.dependencies import get_db
 from app.file_upload import file_upload
 from app.oauth2 import get_current_active_user, oauth_2_scheme
 from app.routers.sessions import create_session_for_conference
 from app.schemas.user_schemas import User
-from ..data_ingestion import createVectorDb, file_path_in_files, write_data_to_json
+from ..data_ingestion import file_path_in_files, write_data_to_json
 from ..data_query import query_document
 from ..crud import conferences_crud, attendee_crud
 from sqlalchemy.orm import Session
-from ..AI_assitant import upload_file, delete_file
+from ..AI_assitant import update_assistant, upload_file, delete_file
 
 router = APIRouter(tags=["ai_models"])
 
 @router.post("/query_document")
 async def query_document_endpoint(question: str, attendee_id:str, conference_id: str, db: Session = Depends(get_db)):
-    assistant_id = conferences_crud.get_conference_by_conference_uuid(db, conference_id).assistant_id
-    thread_id = attendee_crud.get_attendee_by_uuid(db, attendee_id).thread_id
-    file_ids= conferences_crud.get_file_ids_by_conference_id(db, conference_id)
+    conference = conferences_crud.get_conference_by_conference_uuid(db, conference_id)
+    if not conference:
+        raise HTTPException(status_code=404, detail="Conference not found")
+    assistant_id = conference.assistant_id
+    attendee = attendee_crud.get_attendee_by_uuid(db, attendee_id)
+    if not attendee:
+        raise HTTPException(status_code=404, detail="Attendee not found")
+    thread_id = attendee.thread_id
+    file_ids = conferences_crud.get_file_ids_by_conference_id(db, conference_id)
+    if not file_ids:
+        raise HTTPException(status_code=404, detail="No files found for this conference")
+    print(assistant_id,thread_id, file_ids)
     answer = query_document(question,assistant_id,thread_id, file_ids)
     return {"answer": answer}
+    # return {"answer": "answer"}
 
 @router.post("/refresh_input_file/")
 async def update_conference(conference_id: str, current_user: User = Depends(get_current_active_user), db: Session = Depends(get_db)):
     write_data_to_json(conference_id,db)
     file_ids = conferences_crud.get_file_ids_by_conference_id(db, conference_id)
+    if not file_ids:
+        raise HTTPException(status_code=404, detail="No files found for this conference")
     file_id = file_ids[0]
     # delete from openai assistant api
     try:
         delete_file(file_id=file_id)
     except Exception as e:
-        return HTTPException(status_code=400, detail=str(e))
+        raise HTTPException(status_code=400, detail=str(e))
     conferences_crud.delete_file_id(db=db, file_id=file_id, conference_id=conference_id, owner_id=current_user.id)
     # upload it into openai assistant api
     file_path = os.path.join('app', 'files')
@@ -43,9 +56,18 @@ async def update_conference(conference_id: str, current_user: User = Depends(get
         try:
             file = upload_file(file)
         except Exception as e:
-            HTTPException(status_code=400, detail=str(e))
-    conferences_crud.upload_file_id(db=db, file_id = file.id, conference_id=conference_id, owner_id=current_user.id)
-    return {"success": "Conference updated successfully."}
+            raise HTTPException(status_code=400, detail=str(e))
+    try:
+        conferences_crud.upload_file_id(db=db, file_id=file.id, conference_id=conference_id, owner_id=current_user.id)
+    except Exception as e:
+        raise HTTPException(status_code=400, detail=f"Error uploading file: {str(e)}")
+    try:
+        assistant_id = conferences_crud.get_assistant_id_by_conference_id(db, conference_id)
+    except Exception as e:
+        raise HTTPException(status_code=400, detail=f"Error getting assistant ID: {str(e)}")
+    assistant_schema = create_assistant_schema(assistant_id, conference_id, file_id)
+    update_assistant(assistant_schema)
+    return {"success": "Sessions file updated successfully."}
 
 @router.post("/upload_session_file/")
 async def upload_session_file(file: UploadFile,
@@ -58,11 +80,16 @@ async def upload_session_file(file: UploadFile,
     # read the file and return a csv reader object
     reader = await read_file_return_csv(contents,filename)
     headers = reader.fieldnames
+    reader = [{k.lower(): v for k, v in row.items()} for row in reader]
     my_headers = ['name', 'description', 'location', 'date', 'start_time', 'end_time', 'tags', 'speakers']
     if set(headers) != set(my_headers):
-        return {"error": "The attributes(Column Names) provided are not correct.", "expected attributes(Column Names)": my_headers, "received attributes(Column Names)": headers},  
+        error_message = {
+            "error": "The attributes(Column Names) provided are not correct.", 
+            "expected attributes(Column Names)": my_headers, 
+            "received attributes(Column Names)": headers
+        }
+        raise HTTPException(status_code=400, detail=error_message)  
     for row in reader:
-        # print("speakers before conversion", row['speakers'])
         payload = {
             "name": f"{row['name']}" if row['name'] else None,
             "start_time": f"{row['start_time']}" if row['start_time'] else None,
@@ -74,7 +101,10 @@ async def upload_session_file(file: UploadFile,
             "speakers": ast.literal_eval(row['speakers']) if row['speakers'] else None,
             "tags": ast.literal_eval(row['tags']) if row['tags'] else None
         }
-        session = schemas.SessionCreate(**payload)
+        try:
+            session = schemas.SessionCreate(**payload)
+        except Exception as e:
+            raise HTTPException(status_code=400, detail=str(e)+"\n"+str(payload))
         create_session_for_conference(session,db,current_user)
     write_data_to_json(conference_id,db)
     # get the file from the files folder
@@ -84,6 +114,29 @@ async def upload_session_file(file: UploadFile,
         try:
             file = upload_file(file)
         except Exception as e:
-            HTTPException(status_code=400, detail=str(e))
-    conferences_crud.upload_file_id(db=db, file_id = file.id, conference_id=conference_id, owner_id=current_user.id)
+            raise HTTPException(status_code=400, detail=str(e))
+    try:
+        conferences_crud.upload_file_id(db=db, file_id=file.id, conference_id=conference_id, owner_id=current_user.id)
+    except Exception as e:
+        raise HTTPException(status_code=400, detail=f"Error uploading file: {str(e)}")
+    try:
+        assistant_id = conferences_crud.get_assistant_id_by_conference_id(db, conference_id)
+    except Exception as e:
+        raise HTTPException(status_code=400, detail=f"Error getting assistant ID: {str(e)}")
+    assistant_schema = create_assistant_schema(assistant_id, conference_id, file.id)
+    update_assistant(assistant_schema)
     return {'filename':file.filename}
+
+def create_assistant_schema(assistant_id, conference_id, file_id):
+    payload = {
+        "assistant_id": f"{assistant_id}",
+        "model": "gpt-4-1106-preview",
+        "name": f"{conference_id}",
+        "description": "It's a conference assistant. it can help users with their queries related to the sessions of the conference to build their agenda/schedule.",
+        "instructions": "You are conference assistant. You can help users with their queries related to the sessions of the conference to build their agenda/schedule.",
+        "tools": [{"type": "code_interpreter"}],
+        "file_ids": [f"{file_id}"],
+        "metadata": {}
+    }
+    assistant = ai_schemas.AssistantUpdate(**payload)
+    return assistant
